requirements.txt
argh==0.26.2
asyncio==3.4.3
attrs==18.1.0
backcall==0.1.0
decorator==4.3.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pynput==1.4
PyQt5==5.11.2
PyQt5-sip==4.19.12
python-libxdo==0.1.2a1
python-xlib==0.23
PyYAML==3.13
Quamash==0.6.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
watchdog==0.8.3
wcwidth==0.1.7
